refactoring @games.h
