+++ NetWorld-Immersion/Inventory.h
@@ -1,0 +1,6 @@
+//!!! THIS MAUST BE REWORK&CHANGED  !!!
+
+void ShowInventory() {
+    ShowBitmap(0, 0, window.width / 2., window.height / 2., InventoryhBack);//задний фон
+    bool BootsInventory = Boots__inventory_butt.ShowInv();
+}
